Update ScalaDatabase.md
It's not clear how to setup properties for the default pool HikariCP. The 2.4  doc says to inspect the reference.conf, but it doesn't give any examples how to use that information. Also, the standard db.default.url "jdbc:mysql" property is not working in 2.4, I'll look into submitting a bug for that. But the side affect is that the db.default.hikaricp properties in the application.conf aren't recognized if this format of url is used. So it makes one think that they are doing something wrong with the hikaricp properties. So in my example I used the specialized url format "mysql://". One of the big stumbling blocks for our team (being new play users for the last year) has been the lack of completeness in documentation. In general there is a lot of assumed knowledge. I think this addition of an example could help newer Play users.